Added validations to posts controller and Added tests

+++ src/controllers/posts.ts
@@ -1,6 +1,10 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 import IPostService from "../interfaces/posts.interface";
 import { User } from "@prisma/client";
+import "express-async-errors";
+import { nextTick } from "process";
+
+const POST_MIN_LENGTH = 30;
 
 export default class PostController {
   private postService: IPostService;
@@ -15,8 +19,12 @@ export default class PostController {
   };
 
   getPost = async (req: Request, res: Response) => {
-    const id = req.params.id;
+    const { id } = req.params;
+    if (isNaN(parseInt(id))) throw Error("Not Found");
+
     const post = await this.postService.getPost(parseInt(id));
+
+    if (!post) throw Error("Not Found");
     res.render("posts/post", { post: post });
   };
 
@@ -26,15 +34,24 @@ export default class PostController {
 
   createPost = async (req: Request, res: Response) => {
     const { title, content } = req.body;
+
+    if (!title || title.trim() === "") throw Error("Title Missing");
+    if (!content || content.trim().length < POST_MIN_LENGTH) throw Error("Post content is Missing or not long enough");
+
     const user = req.user as User;
     const authorEmail = user.email;
+
     const post = await this.postService.createPost(title, content, authorEmail);
     res.redirect("/posts/" + post.id);
   };
 
   getEditForm = async (req: Request, res: Response) => {
     const { id } = req.params;
+    if (isNaN(parseInt(id))) throw Error("Not Found");
+
     const post = await this.postService.getPost(parseInt(id));
+    if (!post) throw Error("Not Found");
+
     const user = req.user as User;
     if (post?.authorId === user.id) {
       res.render("posts/edit", { post: post });
@@ -44,8 +61,15 @@ export default class PostController {
 
   editPost = async (req: Request, res: Response) => {
     const { id } = req.params;
+    if (isNaN(parseInt(id))) throw Error("Not Found");
+
     const { title, content } = req.body;
+    if (!title || title.trim() === "") throw Error("Title Missing");
+    if (!content || content.trim().length < POST_MIN_LENGTH) throw Error("Post content is Missing or not long enough");
+
     const post = await this.postService.getPost(parseInt(id));
+    if (!post) throw Error("Not Found");
+
     const user = req.user as User;
     if (post?.authorId === user.id) {
       await this.postService.editPost(parseInt(id), title, content);
@@ -55,7 +79,11 @@ export default class PostController {
 
   deletePost = async (req: Request, res: Response) => {
     const { id } = req.params;
+    if (isNaN(parseInt(id))) throw Error("Not Found");
+
     const post = await this.postService.getPost(parseInt(id));
+    if (!post) throw Error("Not Found");
+
     const user = req.user as User;
     if (post?.authorId === user.id) {
       await this.postService.deletePost(parseInt(id));
